move helper function into test file
